Updated WistiaAsset to use an int64 for filesize Updated Wistia direct video URLs to use SSL Fixed a missing async/await in the Campus video request
Former-commit-id: 12967f02df3155da9e29fe2716d4140ed0296da2

+++ church.ccv.CCVRest/Common/Model/WistiaModels.cs
@@ -45,7 +45,7 @@ namespace church.ccv.CCVRest.Common.Model
         public string URL;
         public int Width;
         public int Height;
-        public int FileSize;
+        public UInt64 FileSize;
         public string ContentType;
         public string Type;
     }
